apigw/requirements.txt
# ── Web Framework & ASGI ──────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
hypercorn>=0.16.0
starlette>=0.37.0
gunicorn>=22.0.0

# ── HTTP Client (Async) ───────────────────────────────────────────────────────
httpx>=0.27.0

# ── gRPC ──────────────────────────────────────────────────────────────────────
grpcio>=1.63.0
grpcio-tools>=1.63.0
grpcio-reflection>=1.63.0
protobuf>=4.25.0

# ── WebSocket ─────────────────────────────────────────────────────────────────
websockets>=12.0

# ── Authentication ────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
authlib>=1.3.0
python-multipart>=0.0.9

# ── Rate Limiting ─────────────────────────────────────────────────────────────
slowapi>=0.1.9

# ── Circuit Breaker ───────────────────────────────────────────────────────────
circuitbreaker>=2.0.0

# ── Configuration & Settings ──────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0
pyyaml>=6.0.1

# ── Redis ─────────────────────────────────────────────────────────────────────
redis[asyncio]>=5.0.0

# ── Observability ─────────────────────────────────────────────────────────────
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
prometheus-fastapi-instrumentator>=7.0.0
structlog>=24.1.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.1
tenacity>=8.3.0          # Retry with backoff
cachetools>=5.3.3
aiofiles>=23.2.1
orjson>=3.10.0

# ── Dev / Test ────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-httpx>=0.30.0
httpx>=0.27.0
ruff>=0.4.0
mypy>=1.10.0
